Added "manual override" to get DreamPie out of its stuck state.
The problem seems to be due to the _n_unclaimed_results counter somehow getting out of sync.
At some point, this counter is incremeneted but not properly decremented.
The result is that DreamPie keeps saying "subprocess is busy", thinking it is waiting for a result, but there really isn't any result coming, so it waits forever.
(This waiting occurs in DreamPie.call_subp_noblock, which is called from execute_source.)

The "solution" is to add a simple method to the DreamPie class that just decrements _n_unclaimed_results.
I then added a menu item to call this function.
If this "stuck" situation arises, you can clear it by using this manual override.
Note that using this when the subprocess is *not* stuck could cause havoc, so be careful!